% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic-to-magiceye.R
\name{dino_to_magiceye}
\alias{dino_to_magiceye}
\title{Generate a 3D autostereogram of a dinosaur}
\usage{
dino_to_magiceye(colors = c("#449999", "#FFFFFF", "#010101"), dino = NULL)
}
\arguments{
\item{colors}{Array of at least 2 color hex codes. Colors should be visually distinct and avoid using too many.}

\item{dino}{UID of an image from phylopic.org. See \code{rphylopic} package.}
}
\value{
A raster matrix for \code{plot()}.
}
\description{
Pulls a random dinosaur (or other) image using rphylopic to generate a magic eye.
}
\examples{

 dino_to_magiceye() \%>\% plot()

}
