% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-to-magiceye.R
\name{ggplot_to_magiceye}
\alias{ggplot_to_magiceye}
\title{Convert a ggplot2 object to a 3D autostereogram}
\usage{
ggplot_to_magiceye(
  ggobj,
  colors = c("#00436b", "#ffed89"),
  height = 6,
  width = 6
)
}
\arguments{
\item{ggobj}{A ggplot object.}

\item{colors}{Array of at least 2 color hex codes. Colors should be visually distinct and avoid using too many.}

\item{height}{Height of output matrix. Larger values take longer to render, but will have more detail.}

\item{width}{Width of output matrix. Larger values take longer to render, but will have more detail.}
}
\value{
A raster matrix for \code{plot()}.
}
\description{
Convert a 2ggplot2 object to a 3D autostereogram. Set dot colors used in final image.
}
\examples{
df <- data.frame(
group = c("Dog", "Cat", "Mouse"),
value = c(25, 30, 50)
)

library(ggplot2)
ggplot(df, aes(x="", y=value, fill=group))+
  geom_bar(width = 1, stat = "identity") +
  coord_polar(theta="y") -> pp

pp \%>\% ggplot_to_magiceye() \%>\% plot()
}
