% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magiceye.R
\name{image_to_magiceye}
\alias{image_to_magiceye}
\title{Convert a 2D image to a 3D autostereogram}
\usage{
image_to_magiceye(image, colors = c("#00436b", "#ffed89"), alpha_only = F)
}
\arguments{
\item{image}{Matrix image, usually from \code{png::readPNG()} or \code{jpeg::readJPEG()}.}

\item{colors}{Array of at least 2 color hex codes. Colors should be visually distinct and avoid using too many.}

\item{alpha_only}{Logical. Set to TRUE to only consider alpha channel of a png when rendering image.}
}
\value{
A raster matrix for \code{plot()}.
}
\description{
Convert a 2D greyscale image to a 3D autostereogram. Set dot colors used in final image.
}
\examples{
# Import a jpeg or png
 demo_file <- system.file("extdata", "demo_r.png",
                          package = "rstereogram", mustWork = TRUE)
 demo_image <- png::readPNG(demo_file)

 demo_image \%>\%
    image_to_magiceye() \%>\%
    plot()

}
